+++ SteamTogether.Bot/Services/Command/Factory/ITelegramCommandFactory.cs
@@ -1,4 +1,6 @@
-﻿namespace SteamTogether.Bot.Services.Command.Factory;
+﻿using SteamTogether.Bot.Services.Command.Commands;
+
+namespace SteamTogether.Bot.Services.Command.Factory;
 
 public interface ITelegramCommandFactory
 {
