+++ SteamTogether.Bot/Services/Command/Handlers/TelegramCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using SteamTogether.Bot.Exceptions;
+using SteamTogether.Bot.Services.Command.Commands;
 using SteamTogether.Bot.Services.Command.Factory;
 
 namespace SteamTogether.Bot.Services.Command.Handlers;
